separate octree / genmesh job generic parameterized density function coroutine repeated generation
