--- Packages/com.tuckbone.DualContouring/DualContouring/density.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a3de147ba8f742b4fa317f8cbaa7209a
